Fixed #116 -- Completed initial Provider and Endpoint Definition specs.
* Simplified the database schema. The two major backend collections in
  the current structure are Providers and Signals.
* Converted `main_fixtures.json` into a folder `main_fixtures`. This
  folder contains separate JSON files for each Provider specification.
  The specifications detail the Provider definition and a list of its
  endpoints. The `insert_main_fixtures` management command has been
  updated to iterate over each specification in this folder.
* Added support for more Providers and Endpoints.
  * Twitter - Tweets, Direct Messages Sent/Received
  * Instagram - Posts, Likes
  * Steam - Games Played, Achievements
  * Dropbox - Edits
  * Google - YouTube views, Drive edits, Gmail sent/received
* Updated `scheduleMapper` to actually work on non-trivial Event
  sources.

+++ ografy/core/views/signals.py
@@ -1,5 +1,5 @@
+import datetime
 import json
-import urllib
 
 from django.contrib.auth.decorators import login_required
 from django.core.urlresolvers import reverse
@@ -7,7 +7,7 @@ from django.shortcuts import HttpResponse, HttpResponseRedirect, render
 from mongoengine import Q
 from social.apps.django_app.default.models import UserSocialAuth
 
-from ografy.core.api import EndpointApi, PermissionApi, ProviderApi, SignalApi
+from ografy.core.api import ProviderApi, SignalApi
 from ografy.core.documents import Permission
 
 
@@ -32,8 +32,17 @@ def authorize(request):
         for signal in unverified_signals:
             SignalApi.delete(val=signal.id)
 
+        user_signals = list(SignalApi.get(val=Q(user_id=request.user.id)))
+
         for backend in unassociated_backends:
-            UserSocialAuth.objects.get(id=backend.id).delete()
+            found = False
+
+            for signal in user_signals:
+                if signal.usa_id == backend.id:
+                    found = True
+
+            if not found:
+                UserSocialAuth.objects.get(id=backend.id).delete()
 
         return HttpResponseRedirect(reverse('core_providers'))
 
@@ -75,7 +84,7 @@ def providers(request):
     # FIXME: Make the count happen in the DB
     for provider in providers:
         for signal in signals:
-            if provider.id == signal.provider.id and signal.complete:
+            if provider.provider_number == signal.provider.provider_number and signal.complete:
                 provider.associated_signal = True
 
                 if hasattr(provider, 'assoc_count'):
@@ -94,22 +103,9 @@ def providers(request):
 def verify(request, pk):
     if request.method == 'GET':
         signal = SignalApi.get(Q(user_id=request.user.id) & Q(id=pk)).get()
-        endpoints = EndpointApi.get(Q(provider=signal.provider.id))
-        permissions = PermissionApi.get(Q(signal=signal.id))
-        initial_verification = False
-        endpoint_list = {}
+        event_sources = ProviderApi.get(Q(provider_number=signal.provider.provider_number)).get().event_sources
 
-        for endpoint in endpoints:
-            matched_auth_endpoint = False
-
-            for permission in permissions:
-                if permission.endpoint.id == endpoint.id:
-                    matched_auth_endpoint = True
-                    endpoint_list[endpoint] = permission
-
-            if not matched_auth_endpoint:
-                endpoint_list[endpoint] = None
-
+        # If something went wrong with authorization,
         if signal.connected is False:
             return render(request, 'core/signals/authorize.html', {
                 'title': 'Ografy - Authorize ' + signal.provider.name + ' Connection',  # Change to signal
@@ -119,18 +115,20 @@ def verify(request, pk):
                 'signal': signal
             })
         else:
-            if not signal.complete:
-                initial_verification = True
-                extra_data = UserSocialAuth.objects.filter(user=request.user.id, id=signal.usa_id)[0].extra_data
+            signal_data = UserSocialAuth.objects.filter(user=request.user.id, id=signal.usa_id).get().extra_data
+            # OAuth1 returns tokens on extra_data.  Signal.signal_data is serialized and sent to the user, and we don't
+            # want those tokens available on the client, so delete them from signal.signal_data
+            if signal.provider.auth_type == 1:
+                signal.oauth_token = signal_data['access_token'].pop('oauth_token')
+                signal.oauth_token_secret = signal_data['access_token'].pop('oauth_token_secret')
+            elif signal.provider.auth_type == 0:
+                signal.access_token = signal_data.pop('access_token')
 
-                if signal.provider.auth_type == 1:
-                    access_token = extra_data['access_token']
-                    signal.oauth_token = access_token['oauth_token']
-                    signal.oauth_token_secret = access_token['oauth_token_secret']
-                elif signal.provider.auth_type == 0:
-                    signal.access_token = extra_data['access_token']
+                if hasattr(signal_data, 'refresh_token'):
+                    signal.refresh_token = signal_data.pop('refresh_token')
 
-                signal.save()
+            signal.signal_data = signal_data
+            signal.save()
 
             return render(request, 'core/signals/verify.html', {
                 'title': 'Ografy - Verify ' + signal.provider.name + ' Connection',  # Change to signal
@@ -138,91 +136,46 @@ def verify(request, pk):
                 'content_class': 'left',
                 'signal': signal,
                 'provider': signal.provider,
-                'initial_verification': initial_verification,
                 'current': 'verify',
-                'endpoint_list': endpoint_list
+                'event_source_dict': event_sources
             })
     elif request.method == 'POST':
         signal = SignalApi.get(Q(user_id=request.user.id) & Q(id=pk)).get()
+        provider = ProviderApi.get(Q(provider_number=signal.provider.provider_number)).get()
 
-        if (not signal.complete):
+        if not signal.complete:
             signal.complete = True
             signal.enabled = True
 
-        user_social_auth_instance = UserSocialAuth.objects.filter(id=signal.usa_id).get()
         signal.name = request.POST['name']
         signal.frequency = int(request.POST['updateFrequency'])
-        extra_data = user_social_auth_instance.extra_data
-
-        # OAuth1 returns tokens on extra_data.  Signal.extra_data is serialized and sent to the user, and we don't
-        # want those tokens available on the client, so delete them from signal.extra_data
-        if 'access_token' in extra_data:
-            if 'oauth_token' in extra_data['access_token']:
-                extra_data['access_token'].pop('oauth_token')
-                extra_data['access_token'].pop('oauth_token_secret')
-            else:
-                extra_data.pop('access_token')
-
-        signal.extra_data = extra_data
         signal.save()
 
-        # This handles updating the Permissions enabled status and the creation of Permissions.
-        # The client passes a dictionary of Endpoint Definitions and their associated Permissions to the server.
-        # The dictionary keys are the IDs of the Endpoint Definitions associated with this signal.
-        # The dictionary values are either 'None' if an Permission has not been created for that endpoint
-        # definition yet, or a dictionary if there is an Permission.
-        # The sub-dictionary's key is the ID of the Permission, and the value is a boolean representing
-        # whether or not the Permission is enabled.
-        # Example:
-        # {
-        #   38920743: None,
-        #   32957207: {
-        #       38828502: False
-        #   }
-        # }
-        # Endpoint definition 38920743 has no Permission associated with it.
-        # Endpoint definition 32957207 has Permission 38828502 associated with it.
-        # Permission 38828502 should be disabled.
-        #
+        # Get the event_source dictionary from the request
+        permissions_list = json.loads(request.POST['permissions'])
+        event_sources = provider.event_sources
 
-        # Get the endpoint dictionary from the request
-        endpoints_dict = json.loads(request.POST['endpointsDict'])
-        for index in endpoints_dict:
-            # Get the Permission dictionary for this Endpoint Definition (or None if there is no Permission).
-            this_endpoint_dict = endpoints_dict[index]
-            # Get the Permission ID (or None if it doesn't exist)
-            this_permission_id = list(this_endpoint_dict.keys())[0]
-
-            # If the Permission exists, update its enabled status based on what was passed to the server
-            if not this_permission_id == 'None':
-                PermissionApi.patch(this_permission_id, {
-                    'enabled': this_endpoint_dict[this_permission_id]
-                })
-            else:
-                # If the Permission does not exist and the user wants that endpoint to be used, create an Permission.
-                if this_endpoint_dict[this_permission_id]:
-                    this_endpoint = EndpointApi.get(Q(id=index)).get()
-                    url_parts = [''] * 6
-                    url_parts[0] = this_endpoint.provider.scheme
-                    url_parts[1] = this_endpoint.provider.domain
-                    url_parts[2] = this_endpoint.path
-
-                    if this_endpoint.provider.backend_name == 'facebook':
-                        url_parts[3] = signal.usa_id
-                        # This was the old way of joining, leaving in for reference until this is locked
-                        # route = os.path.join(os.path.join(this_endpoint.provider.base_route, signal.usa_id, this_endpoint.path))
-
-                    route = urllib.parse.urlunparse(url_parts)
-
+        for event_source_name in permissions_list:
+            for event_source in event_sources:
+                if event_source == event_source_name:
                     new_permission = Permission(
-                        name=this_endpoint.name,
-                        route=route,
-                        provider=this_endpoint.provider,
-                        user_id=request.user.id,
-                        signal=signal,
-                        endpoint=this_endpoint,
+                        event_source=event_sources[event_source],
                         enabled=True
                     )
-                    PermissionApi.post(new_permission)
+                    signal.permissions[event_source_name] = new_permission
+                    signal.endpoint_data[event_source_name] = {}
+
+                    for endpoint in event_sources[event_source]['endpoints']:
+                        signal.endpoint_data[event_source_name][endpoint] = {}
+                        parameter_descriptions = signal.permissions[event_source_name]['event_source']['endpoints'][endpoint]['parameter_descriptions']
+
+                        for parameter in signal.permissions[event_source_name]['event_source']['endpoints'][endpoint]['parameter_descriptions']:
+                            if 'default' in parameter_descriptions[parameter].keys():
+                                if parameter_descriptions[parameter]['default'] == 'date_now':
+                                    signal.endpoint_data[event_source_name][endpoint][parameter] = datetime.date.today().isoformat().replace('-', '/')
+                                else:
+                                    signal.endpoint_data[event_source_name][endpoint][parameter] = parameter_descriptions[parameter]['default']
+
+        signal.save()
 
         return HttpResponse(reverse('core_providers'))
